Fixed conditional_installer variables
The policy looks for these variables in 'conditional_installer:main', not
'conditional_installer:conditional_installer'. And the policy looks for 'packages',
not 'name' as the key holing the comma separated list of packages that should be
installed for a given condition.

Ticket: ENT-10745
Changelog: Title

Fixed key name of packages for conditional_installer installed packages