Added modules for autorun-inputs and autorun-bundles
This simply mirrors the existing autorun bundle, but enables inputs and bundles
independently. This is useful in the context of migrating a policy set where you
have a mix of autorun policy that is present in services/autorun as well as
outside that directory.

This enables you to add your custom policy file tagged for autorun outside
services/autorun and have it depend on autorun-bundles.